refactor: move useList to own folder
