+++ packages/dm-core/src/hooks/index.tsx
@@ -5,4 +5,4 @@ export * from './useLocalStorage'
 export * from './useDataSources'
 export * from './useSearch'
 export * from './useRecipe'
-export * from './useList'
+export * from './useList/useList'
